% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blowflydata}
\alias{blowflydata}
\title{Nicholson's Blowflies Data.}
\format{A data frame with two variables: \code{day} and \code{y}.
\code{day} indicates the time observations are made. The experiment last from the 40th day to the 315th day. Observations are made daily.
\code{y} is the blowfly population count.}
\usage{
blowflydata
}
\description{
A set of control experiments in which the adult food supply was restricted to 0.4g per day. The adult blowfly population exhibits oscillation due to constant resource level.
}
\keyword{datasets}

