% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DDEdiag.R
\name{DDEdiag}
\alias{DDEdiag}
\title{Time Series Diagnostics for the Residuals}
\usage{
DDEdiag(y, times, fitted, use.TSA = FALSE, ...)
}
\arguments{
\item{y}{Matrix of observed data values.}

\item{times}{Vector observation times for the data.}

\item{fitted}{The functional data object for the estimated state process.}

\item{use.TSA}{\code{TURE} or \code{FALSE}. Whether to use TSA package to perfor the diagnostics.}

\item{...}{Additional arguments for \code{tsdiag} function.}
}
\value{
None. Diagnostics are plotted.
}
\description{
A function to plot time-series diagnostics of the residuals from a generalized profiling DDE model.
}
\author{
Ziqian Zhou
}
\seealso{
\code{tsdiag}
}

