% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSIR.fit}
\alias{DSIR.fit}
\title{Fitted Delay SIR model}
\format{A list returned by \code{Profile.LS.DDE}}
\usage{
DSIR.fit
}
\description{
Fitted results of the Delay SIR model with data of the \code{DSIR.data} and fitted using the code in the example section of \code{Profile.LS.DDE}.
}
\keyword{datasets}

