% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSIRdata}
\alias{DSIRdata}
\title{Simulated dataset from a delay SIR model}
\format{A matrix of two columns. The first column is the observed state \code{S} and the second column is the observed state \code{I}.}
\usage{
DSIRdata
}
\description{
A simulated dataset from a delay SIR model specified as following:
\deqn{\dot{S} = \rho(t)-\beta I_{\tau}S\\
 \dot{I} = \beta I_{\tau}S-\gamma I}
The parameters are set at: \eqn{\rho(t)=4000\times(\sin(t)+2)}, \eqn{\gamma = 5},
\eqn{\beta = 0.0012}, and \eqn{\tau = 0.5}.  We simulated the numerical solution from time \eqn{t=0} to \eqn{t=50} and the process is then sampled regularly ten times per unit time.  Independent normal observational noise with sd = 100 is added to the numerical solution.
}
\keyword{datasets}

