% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.DSIR.R
\name{DSIRfn.make}
\alias{DSIRfn.make}
\title{Make DTVSIR functions}
\usage{
DSIRfn.make()
}
\value{
A list of functions that calculate the derivatives of delay SIR model with time varying coefficient.
}
\description{
Make DTVSIR functions
}
\examples{
DSIRfn.make()
 DSIRfn <- DSIRfn.make()
}

