% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSIRInitCoefs}
\alias{DSIRInitCoefs}
\title{Initial values for the spline coefficients for the Delay SIR model}
\format{A matrix. Two columns for state "S" and "I" respectively.}
\usage{
DSIRInitCoefs
}
\description{
Initial values for the spline coefficients for the Delay SIR model used in the example of \code{Profile.LS.DDE} with data of the \code{DSIR.data}.
}
\keyword{datasets}

