% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.DSIR.R
\name{DTVSIRfn.make}
\alias{DTVSIRfn.make}
\title{Make DTVSIR functions}
\usage{
DTVSIRfn.make()
}
\value{
A list of functions that calculate the derivatives of delay SIR model with time varying coefficient.
}
\description{
Make DTVSIR functions
}
\examples{
DTVSIRfn.make()
DTVSIRfn <- DTVSIRfn.make()
}

