% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IntegrateForward.DDE.R
\name{IntegrateForward.DDE}
\alias{IntegrateForward.DDE}
\title{Intergrate Forward a DDE Model}
\usage{
IntegrateForward.DDE(times.forecast, pars, beta, proc, more, tau, ndelay,
  fdobj0, fdobj.d)
}
\arguments{
\item{times.forecast}{A time series at which the state of the process is predicted}

\item{pars}{Estimated parameters.}

\item{beta}{Estimated contributions of lags of delay.}

\item{proc}{The \code{proc} object returned from estimation functions.}

\item{more}{An object specifying additional arguments to fn.}

\item{tau}{A list of delay lags.}

\item{ndelay}{A vector inidicating which state process has a delay term.}

\item{fdobj0}{A functional data object fitted with the initial history part of the data.}

\item{fdobj.d}{A functional data object fitted by generalized profiling.}
}
\value{
A list of objects
\describe{
\item{times}{The time points at where the predictions are made.}
\item{states}{The predicted states of the process.}
}
}
\description{
Solves a delay differential equation going forward based on a proc object.
}
\author{
Ziqian Zhou
}

