% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vector.D.funcs.R
\name{make.vector.disease.fn}
\alias{make.vector.disease.fn}
\title{Make Vector Disease functions}
\usage{
make.vector.disease.fn()
}
\value{
A list of functions that calculate the derivatives of vector disease model.
}
\description{
Make Vector Disease functions
}
\examples{
make.vector.disease.fn()
vector.disease.fn <- make.vector.disease.fn()
}

