% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tv-delay-cov.R
\name{ProfileSSE.covariance.DDE}
\alias{ProfileSSE.covariance.DDE}
\title{ProfileSSE.covariance.DDE}
\usage{
ProfileSSE.covariance.DDE(pars, beta, active = NULL, eps = 1e-06, ...)
}
\arguments{
\item{pars}{The estimated parameters.}

\item{beta}{The estimated parameters.}

\item{active}{Incides indicating which parameters of pars should be estimated; defaults to all of them.}

\item{eps}{Step-size for finite difference estimate of second derivatives.}

\item{...}{Additional arguments used for profiling estimation}
}
\value{
Returns a Newey-West estimate of the covariance matrix of the parameter estimates.
}
\description{
Newey-West estimate of covariance of parameter estimates from profiling for DDE models.
Currently assumes a lag-5 auto-correlation among observation vectors.
}
\author{
Ziqian Zhou
}
\seealso{
\code{\link{Profile.LS.DDE}}
}

